requirements.txt
playwright
playwright-stealth
openai
sqlalchemy
python-dotenv
streamlit
pytest
